+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 2.2.3
 
+## New Components
+
+* `ButtonToggleGroup` is a component that allows a group of ButtonToggles to behave as a single form component with label, field help, validation message, warning message, and info message support.
+
 * Fix table-ajax to only set data-state to loaded once all of the data has been set. This resolves an issue when automating user scenarios allowing us to reliabliy wait for the loaded state to be set, before moving on.
 
 # 2.2.2